requirements.txt
langchain
gradio
langchain-ollama
langchain_core
langchain_openai
python-dotenv
langchain-huggingface
transformers
pypdf
langchain_community
bs4